requirements.txt
aliyun-python-sdk-core==2.13.13
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.9.0
amqp==2.6.0
appnope==0.1.2
asgiref==3.5.0
backcall==0.1.0
billiard==3.5.0.5
black==23.1.0
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
click==8.1.3
coreschema==0.0.4
crcmod==1.7
cryptography==36.0.1
decorator==4.4.1
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==2.2.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-reversion==5.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
environs==9.5.0
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==1.7.0
install==1.3.5
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.11
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.2.2
mypy-extensions==1.0.0
mysqlclient==2.1.0
odfpy==1.4.0
openpyxl==3.0.2
oss2==2.9.1
packaging==23.0
parso==0.6.2
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.3
psutil==5.9.4
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.9.4
Pygments==2.5.2
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
qrcode==7.3.1
redis==3.3.11
redislite==6.2.805324
requests==2.22.0
rsa==4.8
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
tomli==2.0.1
traitlets==4.3.3
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.20
vine==1.3.0
wcwidth==0.1.8
Werkzeug==2.2.3
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.1.0
